Move AwaitThread to core
This patch moves the AwaitThread class to core to reuse it in the backup subpackage.
